fix: Исправления после первого ревью Убрала force unwrapping и перенесла всю логику из DataProvider в TrackerStore
