+++ Tracker/Core Data/CoreDataErrors.swift
@@ -1,9 +1,10 @@
 import Foundation
 
-enum DataProviderErrors: Error {
+enum CoreDataErrors: Error {
     case objectNotFound
     case noSectionsAvailable
     case sectionOutOfRange(index: Int)
     case rowOutOfRange(index: Int)
     case trackerConversionError
+    case nilResult
 }
